Standardize API naming conventions
The PECI API had an inconsistent naming convention, which could get
confusing esp. regarding the word "config" (sometimes it means PCI
config space, and sometimes not). This commit renames all the functions
to have consistent naming and capitalization, and renames all the
EndPointConfig function to just start with EndPoint. (It's a problem
with the PECI spec itself that EndPointConfig commands also are used for
MMIO access - but we can write a clearer API ourselves.)

Preprocessor definitions are added so that existing uses of the
functions are simply rewritten to the new function name.

Additionally, all output data parameters are changed to be void* type
rather than uint8_t* type. This makes it easier to use since both C and
C++ allow for implicit conversions from any type* to void*, reducing the
need for explicit casts.

Signed-off-by: Jonathan Doman <jonathan.doman@intel.com>
Change-Id: Id99e999d5b681c8bacedad0927c8405f4aa851b5