Add support for the PECI domain ID byte
The PECI driver has added support for the PECI domain ID byte by adding
it to the existing command structs.

This creates a domain-supported version of the PECI commands that adds
a parameter to set the domain_id in the command struct.

It also updates the existing functions to call the domain-supported
version with a default domain of 0 for backward compatibility.

Tested:
Confirmed that the PECI host ID byte is correctly passed to the
driver when the domain-supported APIs are used.

Signed-off-by: Jason M. Bills <jason.m.bills@intel.com>
Change-Id: If58e65d80a9ad60ff82b9fc023f83ff466e8ea8f