Apply modern compiler options
These are options used in other places that significantly reduce the
size of binaries, in this case, 6KB, or about 15% of the total
application size.

Also, it should be noted, because raw-peci is a c++ application, we need
to set the CMAK_CXX_FLAGS as well for them to be applied.  Specific
flags enabled are in the patchset.

Tested: Code compiles

Signed-off-by: Ed Tanous <edtanous@google.com>
Change-Id: I68c91fa6fede6ab9ae69c6f29a2537ee95584e0a